Whey only for protein: fractional scoops, not a calorie filler; more snacks
Matt's protein was landing at ~246 g against a 180 g target because the
solver pinned whey at the 6-scoop max every day — it was using whey to
fill his calorie *floor* (6 x 95 = 572 kcal), not because he needed the
protein.

- Whey no longer counts toward calories_daily_min (food must carry the
  floor). It still counts toward the ceiling, so it can't overshoot.
- Scoops are fractional end to end (Numeric column, migration 0011) so the
  solver includes only what's necessary; UI shows e.g. '2.4 scoops (60 g)'.
- max_snacks_per_day 3 -> 6 (cap raised to 8) so real food, not whey, can
  close the calorie gap.

+++ src/meal_planner/db/models.py
@@ -286,7 +286,7 @@ class PlanDayProfile(Base):
     protein_g: Mapped[Decimal | None] = mapped_column(Numeric)
     fat_g: Mapped[Decimal | None] = mapped_column(Numeric)
     carbs_g: Mapped[Decimal | None] = mapped_column(Numeric)
-    whey_scoops: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
+    whey_scoops: Mapped[Decimal] = mapped_column(Numeric, nullable=False, default=0)
 
 
 class PlanDay(Base):
